+++ packages/react-app/src/views/index.js
@@ -1,2 +1,2 @@
 export { default as Debug } from "./Debug";
-export { default as Admin } from "./Admin/Admin";
+export { default as Admin } from "./admin/Admin";
